+++ components/checks/active/path_traversal.rb
@@ -9,7 +9,7 @@
 # Path Traversal check.
 #
 # @author Tasos "Zapotek" Laskos <tasos.laskos@arachni-scanner.com>
-# @version 0.4.2
+# @version 0.4.3
 #
 # @see http://cwe.mitre.org/data/definitions/22.html
 # @see http://www.owasp.org/index.php/Path_Traversal
@@ -39,10 +39,7 @@ class Arachni::Checks::PathTraversal < Arachni::Check::Base
             # Add one more mutation (on the fly) which will include the extension
             # of the original value (if that value was a filename) after a null byte.
             each_mutation: proc do |mutation|
-                next if !mutation.affected_input_value ||
-                    (mutation.is_a?( Arachni::Form ) &&
-                        (mutation.mutation_with_original_values? ||
-                            mutation.mutation_with_sample_values?))
+                next if !mutation.affected_input_value
 
                 # Don't bother if the current element type can't carry nulls.
                 next if !mutation.valid_input_value_data?( "\0" )
@@ -115,7 +112,7 @@ of relevant content in the HTML responses.
             elements:    [ Element::Form, Element::Link, Element::Cookie,
                            Element::Header, Element::LinkTemplate ],
             author:      'Tasos "Zapotek" Laskos <tasos.laskos@arachni-scanner.com> ',
-            version:     '0.4.2',
+            version:     '0.4.3',
             platforms:   payloads.keys,
 
             issue:       {